merge Dev

+++ core/frontend/src/views/dataset/exportExcel/index.vue
@@ -21,6 +21,20 @@
       />
     </el-tabs>
     <de-btn
+      v-show="activeName === 'SUCCESS' && multipleSelection.length === 0"
+      secondary
+      icon="el-icon-delete"
+      @click="downLoadAll"
+    >{{ $t("data_export.download_all") }}
+    </de-btn>
+    <de-btn
+      v-show="activeName === 'SUCCESS' && multipleSelection.length !== 0"
+      secondary
+      icon="el-icon-delete"
+      @click="downLoadAll"
+    >{{ $t("data_export.download") }}
+    </de-btn>
+    <de-btn
       v-show="multipleSelection.length === 0"
       secondary
       icon="el-icon-delete"
@@ -62,7 +76,7 @@
               <div class="name-content">
                 <div class="fileName">{{ scope.row.fileName }}</div>
                 <div
-                  v-if="activeName==='FAILED'"
+                  v-if="scope.row.exportStatus==='FAILED'"
                   class="failed"
                 >{{ $t("data_export.export_failed") }}</div>
                 <div
@@ -72,11 +86,11 @@
               </div>
             </div>
             <div
-              v-if="activeName==='FAILED'"
+              v-if="scope.row.exportStatus==='FAILED'"
               class="red-line"
             />
             <el-progress
-              v-if="activeName==='IN_PROGRESS'"
+              v-if="scope.row.exportStatus==='IN_PROGRESS'"
               :percentage="+scope.row.exportPogress"
             />
           </template>
@@ -163,6 +177,7 @@ import request from '@/utils/request'
 import { downloadFile, post } from '@/api/dataset/dataset'
 import bus from '@/utils/bus'
 import { Button } from 'element-ui'
+import { runInContext as tableData } from 'lodash'
 export default {
   mixins: [msgCfm],
   data() {
@@ -212,7 +227,7 @@ export default {
       this.timer = setInterval(() => {
         if (this.activeName === 'IN_PROGRESS') {
           post(
-            '/exportCenter/exportTasks/' + this.activeName, {}, true
+            '/exportCenter/exportTasks/' + this.activeName, {}, false
           ).then(
             (res) => {
               this.tabList.forEach(item => {
@@ -286,6 +301,14 @@ export default {
       bus.$emit('data-export-center')
     },
     handleClick() {
+      if (this.activeName === 'ALL') {
+        this.description = this.$t('data_export.no_file')
+      } else if (this.activeName === 'FAILED') {
+        this.description = this.$t('data_export.no_failed_file')
+      } else {
+        this.description = this.$t('data_export.no_task')
+      }
+
       this.tableData = []
       this.drawerLoading = true
       post(
@@ -320,6 +343,39 @@ export default {
         this.drawerLoading = false
       })
     },
+    downLoadAll() {
+      if (this.multipleSelection.length === 0) {
+        this.tableData.forEach(item => {
+          downloadFile(item.id).then((res) => {
+            const blob = new Blob([res], { type: 'application/vnd.ms-excel' })
+            const link = document.createElement('a')
+            link.style.display = 'none'
+            link.href = URL.createObjectURL(blob)
+            link.download = item.fileName // 下载的文件名
+            document.body.appendChild(link)
+            link.click()
+            document.body.removeChild(link)
+          }).finally(() => {
+            this.exportDatasetLoading = false
+          })
+        })
+        return
+      }
+      this.multipleSelection.map((ele) => {
+        downloadFile(ele.id).then((res) => {
+          const blob = new Blob([res], { type: 'application/vnd.ms-excel' })
+          const link = document.createElement('a')
+          link.style.display = 'none'
+          link.href = URL.createObjectURL(blob)
+          link.download = ele.fileName // 下载的文件名
+          document.body.appendChild(link)
+          link.click()
+          document.body.removeChild(link)
+        }).finally(() => {
+          this.exportDatasetLoading = false
+        })
+      })
+    },
     downloadClick(item) {
       downloadFile(item.id).then((res) => {
         const blob = new Blob([res], { type: 'application/vnd.ms-excel' })
@@ -346,15 +402,28 @@ export default {
       )
     },
     deleteField(item) {
-      request({
-        url: '/exportCenter/delete/' + item.id,
-        method: 'get'
-      }).then(
-        (res) => {
-          this.handleClick()
-        }
-      )
-      this.openMessageSuccess('commons.delete_success')
+      this.$confirm(this.$t('data_export.sure_del'), '', {
+        confirmButtonText: this.$t('commons.delete'),
+        cancelButtonText: this.$t('commons.cancel'),
+        cancelButtonClass: 'de-confirm-fail-btn de-confirm-fail-cancel',
+        confirmButtonClass: 'de-confirm-fail-btn de-confirm-fail-confirm',
+        customClass: 'de-confirm de-confirm-fail',
+        iconClass: 'el-icon-warning'
+      })
+        .then(() => {
+          request({
+            url: '/exportCenter/delete/' + item.id,
+            method: 'get'
+          }).then(
+            (res) => {
+              this.openMessageSuccess('commons.delete_success')
+              this.handleClick()
+            }
+          )
+        })
+        .catch(() => {
+          this.$info(this.$t('commons.delete_cancel'))
+        })
     },
     handleSelectionChange(val) {
       this.multipleSelection = val
@@ -369,28 +438,55 @@ export default {
     },
     delAll() {
       if (this.multipleSelection.length === 0) {
-        post(
-          '/exportCenter/deleteAll/' + this.activeName,
-          this.multipleSelection.map((ele) => ele.id),
-          true
-        ).then(
-          (res) => {
-            this.handleClick()
-          }
-        )
-        this.openMessageSuccess('commons.delete_success')
+        this.$confirm(this.$t('data_export.sure_del_all'), '', {
+          confirmButtonText: this.$t('commons.delete'),
+          cancelButtonText: this.$t('commons.cancel'),
+          cancelButtonClass: 'de-confirm-fail-btn de-confirm-fail-cancel',
+          confirmButtonClass: 'de-confirm-fail-btn de-confirm-fail-confirm',
+          customClass: 'de-confirm de-confirm-fail',
+          iconClass: 'el-icon-warning'
+        })
+          .then(() => {
+            post(
+              '/exportCenter/deleteAll/' + this.activeName,
+              this.multipleSelection.map((ele) => ele.id),
+              true
+            ).then(
+              (res) => {
+                this.openMessageSuccess('commons.delete_success')
+                this.handleClick()
+              }
+            )
+          })
+          .catch(() => {
+            this.$info(this.$t('commons.delete_cancel'))
+          })
         return
       }
-      post(
-        '/exportCenter/delete',
-        this.multipleSelection.map((ele) => ele.id),
-        true
-      ).then(
-        (res) => {
-          this.handleClick()
-        }
-      )
-      this.openMessageSuccess('commons.delete_success')
+
+      this.$confirm(this.$t('data_export.sure_del'), '', {
+        confirmButtonText: this.$t('commons.delete'),
+        cancelButtonText: this.$t('commons.cancel'),
+        cancelButtonClass: 'de-confirm-fail-btn de-confirm-fail-cancel',
+        confirmButtonClass: 'de-confirm-fail-btn de-confirm-fail-confirm',
+        customClass: 'de-confirm de-confirm-fail',
+        iconClass: 'el-icon-warning'
+      })
+        .then(() => {
+          post(
+            '/exportCenter/delete',
+            this.multipleSelection.map((ele) => ele.id),
+            true
+          ).then(
+            (res) => {
+              this.openMessageSuccess('commons.delete_success')
+              this.handleClick()
+            }
+          )
+        })
+        .catch(() => {
+          this.$info(this.$t('commons.delete_cancel'))
+        })
     },
 
     handleClose() {
